Refinement Coffee Machine
Program work endlessly to make coffee for all interested persons until the shutdown signal is given.
Presented two new options: "remaining" and "exit".
If the coffee machine doesn't have enough resources to make coffee, the program output a message that says it can't make a cup of coffee.
If the user types "buy" to buy a cup of coffee and then changes his mind, they can to type "back" to return into the main cycle.